Terava/backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.30.0
cryptography==38.0.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==1.0.0
dj-rest-auth==2.2.5
Django==4.1
django-allauth==0.51.0
django-cors-headers==3.13.0
django-extensions==3.2.1
django-taggit==3.0.0
djangorestframework==3.13.1
drf-spectacular==0.24.2
easy-thumbnails==2.8.3
environs==9.5.0
flower==1.2.0
gunicorn==20.1.0
humanize==4.3.0
idna==3.4
inflection==0.5.1
jsonschema==4.16.0
kombu==5.2.4
marshmallow==3.18.0
oauthlib==3.2.1
packaging==21.3
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
toml==0.10.2
tornado==6.2
types-cryptography==3.3.23
uritemplate==4.1.1
urllib3==1.26.12
vercel==0.2.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
